moved LH functions
